Renaming classes and running tests
