--- src/main/java/uk/ac/roe/wfau/enteucha/crossmatch/CrossMatcher.java
@@ -1,33 +1,0 @@
-/*
- *  Copyright (C) 2018 Royal Observatory, University of Edinburgh, UK
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package uk.ac.roe.wfau.enteucha.crossmatch;
-
-import org.springframework.stereotype.Component;
-
-import lombok.extern.slf4j.Slf4j;
-
-/**
- * 
- * 
- */
-@Slf4j
-@Component
-public class CrossMatcher
-    {
-    }
